+++ packages/seed/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @lunora/seed [1.0.0-alpha.78](https://github.com/anolilab/lunora/compare/@lunora/seed@1.0.0-alpha.77...@lunora/seed@1.0.0-alpha.78) (2026-08-23)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.79
+* **@lunora/testing:** upgraded to 1.0.0-alpha.116
+
 ## @lunora/seed [1.0.0-alpha.77](https://github.com/anolilab/lunora/compare/%40lunora%2Fseed%401.0.0-alpha.76...%40lunora%2Fseed%401.0.0-alpha.77) (2026-08-18)